+++ package.json
@@ -19,5 +19,11 @@
   "homepage": "https://github.com/lisandroseia/awsm-books-moduled#readme",
   "dependencies": {
     "luxon": "^2.3.0"
+  },
+  "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.25.4"
   }
 }
